pyproject.toml
[tool.poetry]
name = "leave-management-system"
version = "0.1.0"
description = "Leave Management System (Flask + SQLAlchemy + Alembic)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0"
sqlalchemy = "^2.0"
alembic = "^1.13"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
pyyaml = "^6.0"
gunicorn = "^22.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
ruff = "^0.5.0"
isort = "^5.13.2"
bandit = "^1.7.9"
semgrep = "^1.78.0"
pytest = "^8.2.0"
pytest-cov = "^4.1.0"
pre-commit = "^3.7.0"
yamllint = "^1.35.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# --- Tool configs (keeps linters/formatters consistent) ---
[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "E402"]  # defer to Black for line length, allow imports after code

[tool.pytest.ini_options]
addopts = "-q --disable-warnings --maxfail=1"
testpaths = ["tests"]
